requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
python-socketio[client]>=5.11.0
boto3>=1.34.0
